+++ src/net/strategies/FailoverStrategy.cpp
@@ -140,7 +140,7 @@ void FailoverStrategy::onResultAccepted(Client *client, int64_t seq, uint32_t di
 
 void FailoverStrategy::add(const Url *url, const char *agent)
 {
-    Client *client = new Client(m_pools.size(), agent, this);
+    Client *client = new Client((int) m_pools.size(), agent, this);
     client->setUrl(url);
     client->setRetryPause(Options::i()->retryPause() * 1000);
